bug: Fixing the functions that convert time to human readable format since days was truncated
The functions that can convert a number of seconds to a human readable format (years, days, hours, minutes, seconds) were truncating the days to a uint8.
The number of days in a year is bigger than a uint8 though, so it needed to be changed to a uint16 to get the output correct.
For many drive functions, this does not matter since things will generally take at most a day, maybe a few, well below 256, however power on hours/minutes can easily surpass this, and this was causing an incorrect value to be output for a human readable time once they number of days surpassed 255 in a given year worth of power on time.

This is a fairly big reaching change and breaks the existing API calls, so the version has received a major version number change to reflect this.

+++ include/common.h
@@ -1213,9 +1213,9 @@ extern "C"
     //  Exit:
     //
     //-----------------------------------------------------------------------------
-    void convert_Seconds_To_Displayable_Time(uint64_t secondsToConvert, uint8_t *years, uint8_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
+    void convert_Seconds_To_Displayable_Time(uint64_t secondsToConvert, uint8_t *years, uint16_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
 
-    void convert_Seconds_To_Displayable_Time_Double(double secondsToConvert, uint8_t *years, uint8_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
+    void convert_Seconds_To_Displayable_Time_Double(double secondsToConvert, uint8_t *years, uint16_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
 
     //-----------------------------------------------------------------------------
     //
@@ -1250,7 +1250,7 @@ extern "C"
     //  Exit:
     //
     //-----------------------------------------------------------------------------
-    void print_Time_To_Screen(uint8_t *years, uint8_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
+    void print_Time_To_Screen(uint8_t *years, uint16_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
 
     extern uint32_t seed32Array[2];
     extern uint64_t seed64Array[2];
